Issue #36 https://github.com/flashbots/op-rbuilder/issues/36
Migrate the rest of the test to the new test utility introduced in #606
---

This PR introduces the following changes:
- Migrates all non-flashblocks integration tests to the new test harness framework
- Fixes all build warnings
- Adds few helper functions and types for testing
- Few refactorings of the test infrastructure
- Refactored the way the `tester` binary is linking to op-rbuilder dependencies
- Moved all tests under `tests/` directory
- Removed `integration` directory`

now running `cargo test` will output

```
$ cargo test
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.36s
     Running unittests src/lib.rs (target/debug/deps/op_rbuilder-7bb461ec2888ee4b)

running 8 tests
test tx_signer::test::test_sign_transaction ... ok
test tests::vanilla::smoke::get_payload_close_to_fcu ... ok
test tests::vanilla::smoke::transaction_flood_no_sleep ... ok
test tests::vanilla::ordering::fee_priority_ordering ... ok
test tests::vanilla::revert::monitor_transaction_drops ... ok
test tests::vanilla::revert::revert_protection ... ok
test tests::vanilla::revert::revert_protection_disabled ... ok
test tests::vanilla::smoke::chain_produces_blocks ... ok

test result: ok. 8 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 23.67s

     Running unittests src/main.rs (target/debug/deps/op_rbuilder-87f69711ab7f0a5e)

running 7 tests
test generator::tests::test_block_cell_update_value ... ok
test generator::tests::test_block_cell_immediate_value ... ok
test generator::tests::test_job_deadline ... ok
test tx_signer::test::test_sign_transaction ... ok
test generator::tests::test_block_cell_wait_for_value ... ok
test generator::tests::test_block_cell_multiple_waiters ... ok
test generator::tests::test_payload_generator ... ok

test result: ok. 7 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 4.70s

   Doc-tests op_rbuilder

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

```
