+++ crates/op-rbuilder/src/lib.rs
@@ -1,4 +1,7 @@
-pub mod integration;
+// pub mod integration;
 pub mod primitives;
-pub mod tester;
+// pub mod tester;
 pub mod tx_signer;
+
+#[cfg(any(test, feature = "tests"))]
+pub mod tests;
