+++ crates/op-rbuilder/src/main.rs
@@ -2,28 +2,28 @@ use args::CliExt;
 use clap::Parser;
 use reth_optimism_cli::{chainspec::OpChainSpecParser, Cli};
 use reth_optimism_node::{node::OpAddOnsBuilder, OpNode};
-
-#[cfg(feature = "flashblocks")]
-use payload_builder::CustomOpPayloadBuilder;
-#[cfg(not(feature = "flashblocks"))]
-use payload_builder_vanilla::CustomOpPayloadBuilder;
 use reth_transaction_pool::TransactionPool;
 
 /// CLI argument parsing.
 pub mod args;
 pub mod generator;
-#[cfg(test)]
-mod integration;
 mod metrics;
 mod monitor_tx_pool;
+mod primitives;
+mod tx_signer;
+
 #[cfg(feature = "flashblocks")]
 pub mod payload_builder;
+
 #[cfg(not(feature = "flashblocks"))]
 mod payload_builder_vanilla;
-mod primitives;
-#[cfg(test)]
-mod tester;
-mod tx_signer;
+
+#[cfg(not(feature = "flashblocks"))]
+use payload_builder_vanilla::CustomOpPayloadBuilder;
+
+#[cfg(feature = "flashblocks")]
+use payload_builder::CustomOpPayloadBuilder;
+
 use metrics::{
     VersionInfo, BUILD_PROFILE_NAME, CARGO_PKG_VERSION, VERGEN_BUILD_TIMESTAMP,
     VERGEN_CARGO_FEATURES, VERGEN_CARGO_TARGET_TRIPLE, VERGEN_GIT_SHA,
