+++ crates/op-rbuilder/src/tests/flashblocks/mod.rs
@@ -1,0 +1,1 @@
+mod smoke;
