+++ crates/op-rbuilder/src/tests/framework/artifacts/genesis.json.tmpl
