+++ crates/op-rbuilder/src/tests/mod.rs
@@ -1,0 +1,9 @@
+// base
+mod framework;
+pub use framework::*;
+
+#[cfg(not(feature = "flashblocks"))]
+mod vanilla;
+
+#[cfg(feature = "flashblocks")]
+mod flashblocks;
