+++ crates/op-rbuilder/src/tests/vanilla/mod.rs
@@ -1,0 +1,5 @@
+mod ordering;
+mod revert;
+mod smoke;
+
+use super::*;
